pyproject.toml
[tool.poetry.scripts]
genbook = "genbook.__main__:main"
[tool.poetry]
name = "genbook"
version = "0.1.0"
description = "An automated tool to generate well-structured books effortlessly using LangChain and Google Gen AI SDK."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "*"
google-genai = "*"
requests = "*"
markdown = "^3.7"
ebooklib = "^0.18"
langgraph = "0.6.6"
typer = "^0.10.0"
click = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

